+++ src/matrix/__init__.py
@@ -1,12 +1,10 @@
 """GA Matrix: label-aware matrix utilities."""
 
 from ._version import __version__
-from .LabeledMatrix import LabeledMatrix
 from .matrix_od import LabelsInput, MatrixInit, MatrixOD
 from .matrix_odt import MatrixODT, Timestamp
 
 __all__ = [
-    "LabeledMatrix",
     "MatrixOD",
     "MatrixODT",
     "__version__",